config: primeiras configurações
