+++ package.json
@@ -1,9 +1,17 @@
 {
+  "version": "1.0.0",
+  "scripts": {
+    "start": "tsc && node dist/index.js",
+    "dev": "nodemon --exec ts-node src/index.ts"
+  },
   "dependencies": {
     "@prisma/client": "^5.9.1",
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "prisma": "^5.9.1"
   },
   "devDependencies": {
-    "@types/express": "^4.17.21"
+    "@types/express": "^4.17.21",
+    "nodemon": "^3.0.3",
+    "ts-node": "^10.9.2"
   }
 }
