+++ src/controllers/produtosController.ts
@@ -1,33 +1,33 @@
-// src/controllers/produtosController.ts
-import { Request, Response } from 'express';
-import { PrismaClient } from '@prisma/client';
+// // src/controllers/produtosController.ts
+// import { Request, Response } from 'express';
+// import { PrismaClient } from '@prisma/client';
 
-const prisma = new PrismaClient();
+// const prisma = new PrismaClient();
 
-export async function getProdutos(req: Request, res: Response): Promise<void> {
-    try {
-        const produtos = await prisma.produto.findMany();
-        res.json(produtos);
-    } catch (error) {
-        console.error('Erro ao obter produtos:', error);
-        res.status(500).json({ error: 'Erro ao obter produtos' });
-    }
-}
+// export async function getProdutos(req: Request, res: Response): Promise<void> {
+//     try {
+//         const produtos = await prisma.produto.findMany();
+//         res.json(produtos);
+//     } catch (error) {
+//         console.error('Erro ao obter produtos:', error);
+//         res.status(500).json({ error: 'Erro ao obter produtos' });
+//     }
+// }
 
-export async function criarProduto(req: Request, res: Response): Promise<void> {
-    const { nome, categoria, preco, quantidade_estoque } = req.body;
-    try {
-        const produto = await prisma.produto.create({
-            data: {
-                nome,
-                categoria,
-                preco,
-                quantidade_estoque
-            }
-        });
-        res.json(produto);
-    } catch (error) {
-        console.error('Erro ao criar produto:', error);
-        res.status(500).json({ error: 'Erro ao criar produto' });
-    }
-}
+// export async function criarProduto(req: Request, res: Response): Promise<void> {
+//     const { nome, categoria, preco, quantidade_estoque } = req.body;
+//     try {
+//         const produto = await prisma.produto.create({
+//             data: {
+//                 nome,
+//                 categoria,
+//                 preco,
+//                 quantidade_estoque
+//             }
+//         });
+//         res.json(produto);
+//     } catch (error) {
+//         console.error('Erro ao criar produto:', error);
+//         res.status(500).json({ error: 'Erro ao criar produto' });
+//     }
+// }
